@@ -1,6 +1,6 @@
 import { Component, OnInit } from '@angular/core';
 import { ServerService } from '../server.service';
-import { FormBuilder, Validators } from '@angular/forms';
+import { UntypedFormBuilder, Validators } from '@angular/forms';
 
 const DEFAULT_SERVER = 'ws://localhost:8080';
 
@@ -18,7 +18,7 @@ export class ConnectComponent implements OnInit {
 
   constructor(
     private server: ServerService,
-    private formBuilder: FormBuilder,
+    private formBuilder: UntypedFormBuilder,
   ) {}
 
   ngOnInit(): void {}
